pyproject.toml
[project]
name = "ping-monitor"
version = "0.1.0"
description = "Windows ping monitor with Tkinter tray UI"
requires-python = ">=3.14"
dependencies = [
    "pillow>=10.2.0",
    "psutil>=5.9.8",
    "pystray>=0.19.5",
    "pywin32>=306",
    "ttkbootstrap-icons>=4.0.0",
    "ttkbootstrap-icons-fa>=1.0.2",
    "winshell>=0.6",
]
